pyproject.toml
[project]
name = "calmmage"
version = "2025.1.4" # bugfix open-dir
description = "Personal GitHub repo for tools, projects, and personal website"
readme = "README.md"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
]

authors = [
    {name = "Petr Lavrov", email = "me@petrlavrov.com"},
]
requires-python = "<4,>=3.12"
dependencies = [
    "typer[all]>=0.9.0",
    "tqdm",
    "loguru",
    # "calmlib @ git+https://github.com/calmmage/calmlib.git@main",
    "calmlib @ {root:uri}/src/calmlib",
    #    "botspot @ git+https://github.com/calmmage/botspot.git@main",
    "PyGithub",
    "python-dotenv",
    "gitpython",
    "pillow<11.0.0,>=10.3.0",
    "pytimeparse<2.0.0,>=1.1.8",
    "telethon<2.0.0,>=1.38.1",
    "pymongo>=4.9",
    "matplotlib<4.0.0,>=3.10.0",
    "seaborn<1.0.0,>=0.13.2",
    "boto3<2.0.0,>=1.35.90",
    "deprecated<2.0.0,>=1.2.15",
    "motor<4.0.0,>=3.6.1",
    "langchain<1.0.0,>=0.3.14",
    "langchain-anthropic<1.0.0,>=0.3.3",
    "langchain-openai<1.0.0,>=0.3.0",
    "langchain-google-genai<3.0.0,>=2.0.9",
    "langchain-ollama<1.0.0,>=0.2.2",
    "langchain-community<1.0.0,>=0.3.14",
    "pandas<3.0.0,>=2.2.3",
    "qrcode>=8.0,<9.0",
    "apscheduler<4.0.0,>=3.11.0",
    "markdown<4.0,>=3.8",
    "mistune<4.0.0,>=3.1.3",
    "pyrogram<3.0.0,>=2.0.106",
    #easydict = ">=1.10"
    #fastapi = "*"
    #fastui = "*"
    #pyperclip = "*"
    #pytest = "*"
    #tqdm = "*"
    #matplotlib = "*"
    #numpy = "*"
    #pandas = "*"
    #seaborn = "*"
    #ipykernel = "*"
    #ipywidgets = "*"
    #joblib = "*"
    #retry = "*"
    #pyperclip = "*"
    #paramiko = "*"
    #scp = "*"
    #pymorphy2 = "*"
    #langchain = "*"
    "croniter>=6.0.0",
    "dotenv>=0.9.9",
    "pyperclip>=1.9.0",
]

[project.urls]
Home = "https://github.com/calmmage/calmmage"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["calmmage"]

[dependency-groups]
dev = [
    "pyinstaller",
    "jupyter>=1.0.0",
    "pytest",
    "black>=24",
    "pre-commit",
    "jupyter-contrib-nbextensions",
    "fastapi",
    "fastui",
    "numpy",
    "plotly",
    "pandas<3.0.0,>=2.2.3",
    "google-auth-oauthlib<2.0.0,>=1.2.1",
    "google-api-python-client<3.0.0,>=2.161.0",
    "requests<3.0.0,>=2.32.3",
    "google-auth-httplib2<1.0.0,>=0.2.0",
]

test = [
    "pytest>=6",
    "isort<7.0.0,>=6.0.0",
    "vulture<3.0,>=2.14",
    "pytest-cov<7.0.0,>=6.1.1",
    "pytest-asyncio<2.0.0,>=1.0.0",
    "pytest-mock<4.0.0,>=3.14.1",
    "pyupgrade<4.0.0,>=3.20.0",
    "pyright<2.0.0,>=1.1.401",
    "ruff<1.0.0,>=0.11.12",
]
docs = [
    #mkdocs = ">=1.6"
]
